fix(desktop): keep proxy setup with server env
